+++ src/nchg/filter/filter.cpp
@@ -246,7 +246,7 @@ using FlatPvalueMap = phmap::flat_hash_map<std::size_t, double>;
 }
 
 struct NCHGFilterResult {
-  hictk::Pixel<std::uint32_t> pixel{};
+  hictk::Pixel<std::uint64_t> pixel{};
   double expected{};
   double pval{};
   double pval_corrected{};
